Refactor task update handling and improve validation.
Replaces category-specific DTOs with a unified `TaskUpdateRequest` for better flexibility. Introduces detailed validation for category-specific fields and improves exception handling in task updates. Adds null user handling in task-user update and enhances error responses for task and user deletions.

+++ src/main/java/com/jumpsoft/taskmanagement/service/TaskService.java
@@ -3,10 +3,15 @@ package com.jumpsoft.taskmanagement.service;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.util.NullableUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.function.ThrowingConsumer;
+import org.springframework.validation.BeanPropertyBindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,12 +19,12 @@ import com.jumpsoft.taskmanagement.controller.CustomException;
 import com.jumpsoft.taskmanagement.dto.TaskFilter;
 import com.jumpsoft.taskmanagement.dto.error.ErrorCode;
 import com.jumpsoft.taskmanagement.dto.task.BugTaskCreateRequest;
-import com.jumpsoft.taskmanagement.dto.task.BugTaskUpdateRequest;
 import com.jumpsoft.taskmanagement.dto.task.FeatureTaskCreateRequest;
-import com.jumpsoft.taskmanagement.dto.task.FeatureTaskUpdateRequest;
 import com.jumpsoft.taskmanagement.dto.task.TaskCreateRequest;
 import com.jumpsoft.taskmanagement.dto.task.Task;
 import com.jumpsoft.taskmanagement.dto.task.TaskUpdateRequest;
+import com.jumpsoft.taskmanagement.dto.task.UpdateTaskInvalidArguments;
+import com.jumpsoft.taskmanagement.enums.TaskCategory;
 import com.jumpsoft.taskmanagement.enums.TaskStatus;
 import com.jumpsoft.taskmanagement.mapper.TaskMapper;
 import com.jumpsoft.taskmanagement.entity.Bug;
@@ -27,6 +32,7 @@ import com.jumpsoft.taskmanagement.entity.Feature;
 import com.jumpsoft.taskmanagement.entity.User;
 import com.jumpsoft.taskmanagement.repository.TaskRepository;
 import com.jumpsoft.taskmanagement.repository.UserRepository;
+import com.jumpsoft.taskmanagement.util.TaskSpecification;
 
 import jakarta.persistence.EntityNotFoundException;
 
@@ -46,6 +52,7 @@ public class TaskService {
 
     @Autowired
     private TaskMapper taskMapper;
+
 
     /**
      * Retrieves a task by its ID and converts it to a DTO representation.
@@ -89,6 +96,13 @@ public class TaskService {
                 .toList();
     }
 
+    public TaskCategory determineTaskCategory(Long taskId) throws EntityNotFoundException {
+        com.jumpsoft.taskmanagement.entity.Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new EntityNotFoundException("Task not found with ID: " + taskId));
+
+        return task.getCategory(); // Returns the TaskType enum (BUG or FEATURE)
+    }
+
     /**
      * Creates a new task in the system based on the provided request data.
      *
@@ -123,73 +137,107 @@ public class TaskService {
         return mapToDTO(savedTask);
     }
 
-    /**
-     * Updates an existing task with the provided request data.
+    /**     * Updates an existing task with the provided request data.
      *
-     * @param id          The unique identifier of the task to update.
-     * @param taskRequest The data required to update the task, including optional fields such as name and description.
-     * @return The updated task as a DTO.
+     * @param id                  The unique identifier of the task to update.
+     * @param taskRequest         The data to update the task with, including category-specific fields.
+     * @param invalidRequestTypeConsumer Consumer to handle invalid arguments for specific task categories.
+     * @return The updated task converted into a DTO format.
      * @throws EntityNotFoundException If the task with the given ID does not exist.
-     * @throws CustomException         If the assigned user is not found or validation fails.
+     * @throws CustomException          If the assigned user is not found or if there are validation errors.
+     * @throws MethodArgumentNotValidException If the provided request data is invalid.
      */
     @Transactional
-    public Task updateTask(Long id, TaskUpdateRequest taskRequest) throws EntityNotFoundException, CustomException {
+    public Task updateTask(Long id, TaskUpdateRequest taskRequest, ThrowingConsumer<UpdateTaskInvalidArguments> invalidRequestTypeConsumer) throws EntityNotFoundException, CustomException, MethodArgumentNotValidException {
 
-        return mapToDTO(taskRepository.findById(id).map(existingTask -> {
-           User assignedUser;
-           if (taskRequest.getUserId() != null) {
-               assignedUser = userRepository.findById(taskRequest.getUserId()).orElseThrow(() ->
-                       new RuntimeException(new CustomException(ErrorCode.USER_WITH_ID_NOT_FOUND, taskRequest.getUserId().toString()))
-               );
-               if (existingTask.getUser() != null && !existingTask.getUser().getId().equals(taskRequest.getUserId())) {
-                     existingTask.setUser(assignedUser);
-               }
-           }
-           if (taskRequest.getName() != null) {
-               existingTask.setName(taskRequest.getName());
-           }
-           if (taskRequest.getDescription() != null) {
-               existingTask.setDescription(StringUtils.isEmpty(taskRequest.getDescription()) ? null : taskRequest.getDescription());
-           }
-           if (taskRequest.getStatus() != null) {
-               existingTask.setStatus(taskRequest.getStatus());
-           }
+        try {
+            return mapToDTO(taskRepository.findById(id).map(existingTask -> {
+                //check userId and update user if necessary
+                User assignedUser;
+                if (taskRequest.userId() != null) {
+                    assignedUser = userRepository.findById(taskRequest.userId()).orElseThrow(() ->
+                            new RuntimeException(new CustomException(ErrorCode.USER_WITH_ID_NOT_FOUND, taskRequest.userId().toString()))
+                    );
+                    if (existingTask.getUser() != null && !existingTask.getUser().getId().equals(taskRequest.userId())) {
+                        existingTask.setUser(assignedUser);
+                    }
+                }
+                if (taskRequest.name() != null) {
+                    existingTask.setName(taskRequest.name());
+                }
+                if (taskRequest.description() != null) {
+                    existingTask.setDescription(StringUtils.isEmpty(taskRequest.description()) ? null : taskRequest.description());
+                }
+                if (taskRequest.status() != null) {
+                    existingTask.setStatus(taskRequest.status());
+                }
 
-           switch (existingTask.getCategory()) {
-                case BUG -> {
-                    if(taskRequest instanceof BugTaskUpdateRequest) {
-                        BugTaskUpdateRequest bugRequest = (BugTaskUpdateRequest) taskRequest;
+                //check category-specific fields
+                switch (existingTask.getCategory()) {
+                    case BUG -> {
+                        if (taskRequest.deadline() != null || taskRequest.businessValue() != null) {
+                            List<String> invalidArguments = new ArrayList<>();
+                            if (taskRequest.deadline() != null) {
+                                invalidArguments.add("deadline");
+                            }
+                            if (taskRequest.businessValue() != null) {
+                                invalidArguments.add("businessValue");
+                            }
+                            try {
+                                invalidRequestTypeConsumer.acceptWithException(UpdateTaskInvalidArguments.of(existingTask.getCategory(), invalidArguments));
+                            } catch (Exception e) {
+                                throw new RuntimeException(e);
+                            }
+                        }
                         Bug bugTask = (Bug) existingTask;
-                        if (bugRequest.severity() != null) {
-                            bugTask.setSeverity(bugRequest.severity());
+                        if (taskRequest.severity() != null) {
+                            bugTask.setSeverity(taskRequest.severity());
                         }
-                        if (bugRequest.reproduceSteps() != null) {
-                            bugTask.setStepsToReproduce(StringUtils.isEmpty(bugRequest.reproduceSteps()) ? null : bugRequest.reproduceSteps());
+                        if (taskRequest.reproduceSteps() != null) {
+                            bugTask.setStepsToReproduce(StringUtils.isEmpty(taskRequest.reproduceSteps()) ? null : taskRequest.reproduceSteps());
                         }
-                    } else {
-                        throw new IllegalArgumentException("Invalid request type for BUG category");
                     }
-                }
-                case FEATURE -> {
-                    if (taskRequest instanceof FeatureTaskUpdateRequest) {
-                        FeatureTaskUpdateRequest featureRequest = (FeatureTaskUpdateRequest) taskRequest;
+                    case FEATURE -> {
+                        if (taskRequest.reproduceSteps() != null || taskRequest.severity() != null) {
+                            List<String> invalidArguments = new ArrayList<>();
+                            if (taskRequest.reproduceSteps() != null) {
+                                invalidArguments.add("reproduceSteps");
+                            }
+                            if (taskRequest.severity() != null) {
+                                invalidArguments.add("severity");
+                            }
+                            try {
+                                invalidRequestTypeConsumer.acceptWithException(UpdateTaskInvalidArguments.of(existingTask.getCategory(), invalidArguments));
+                            } catch (Exception e) {
+                                throw new RuntimeException(e);
+                            }
+                        }
                         Feature featureTask = (Feature) existingTask;
-                        if (featureRequest.businessValue() != null) {
-                            featureTask.setBusinessValue(StringUtils.isEmpty(featureRequest.businessValue()) ? null : featureRequest.businessValue());
+                        if (taskRequest.businessValue() != null) {
+                            featureTask.setBusinessValue(StringUtils.isEmpty(taskRequest.businessValue()) ? null : taskRequest.businessValue());
                         }
-                        if (featureRequest.deadline() != null) {
-                            featureTask.setDeadline(featureRequest.deadline());
+                        if (taskRequest.deadline() != null) {
+                            featureTask.setDeadline(taskRequest.deadline());
                         }
-                    } else {
-                        throw new IllegalArgumentException("Invalid request type for FEATURE category");
                     }
-
+                    default ->
+                            throw new IllegalArgumentException("Unknown task category: " + existingTask.getCategory());
                 }
-                default -> throw new IllegalArgumentException("Unknown task category: " + existingTask.getCategory());
+                return taskRepository.save(existingTask);
+            }).orElseThrow(() -> new EntityNotFoundException("Task with id " + id + " not found")));
+        } catch (RuntimeException e) {
+            switch (e.getCause()){
+                case CustomException customException -> {
+                    throw customException;
+                }
+                case MethodArgumentNotValidException methodArgumentNotValidException -> {
+                    throw methodArgumentNotValidException;
+                }
+                default -> {
+                    throw e;
+                }
             }
-            return taskRepository.save(existingTask);
         }
-        ).orElseThrow(() -> new EntityNotFoundException("Task with id " + id + " not found")));
 
     }
 
@@ -201,6 +249,10 @@ public class TaskService {
      */
     @Transactional
     public void deleteTask(Long id) throws EntityNotFoundException {
+        // Check if task exists first
+        if (!taskRepository.existsById(id)) {
+            throw new EntityNotFoundException("Task with id " + id + " not found");
+        }
         taskRepository.deleteById(id);
     }
 
@@ -213,12 +265,15 @@ public class TaskService {
     }
 
     private Specification<com.jumpsoft.taskmanagement.entity.Task> createStatusAndUserSpecification(TaskStatus status, Long userId) {
-        return (root, query, criteriaBuilder) ->
-                criteriaBuilder.and(
-                        criteriaBuilder.equal(root.get("status"), status),
-                        criteriaBuilder.equal(root.get("user").get("id"), userId)
-                );
-    }
 
+        Specification<com.jumpsoft.taskmanagement.entity.Task> spec = Specification.where(null);
+        if (status != null) {
+            spec = spec.and(TaskSpecification.withStatus(status));
+        }
+        if (userId != null) {
+            spec = spec.and(TaskSpecification.withUserId(userId));
+        }
+        return spec;
+    }
 
 }
